requirements.txt
pycodestyle
autopep8
django
djangorestframework
environ